scripts/requirements.txt
numpy
pandas
scikit-learn
scipy
Pillow
tqdm

tensorflow
keras
imbalanced-learn
opencv-python

albumentations
scikit-image

matplotlib
seaborn
plotly
wandb 

astropy
photutils  

joblib
memory_profiler
psutil
pyarrow 

jupyter
jupyterlab
ipython
pdoc3  


shap
lime
eli5


xgboost  
lightgbm
catboost